requirements.txt
pandas==2.1.4
numpy==1.24.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
requests==2.31.0
boto3==1.34.0
apache-airflow==2.8.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
